배너 색 변경, 배너 위치를 레이아웃 안으로 이동
