+++ src/app/layout.tsx
@@ -8,6 +8,7 @@ import { store } from '../store/index';
 import { Providers } from './providers';
 import NavigationBar from '@/components/NavigationBar';
 import Footer from '@/components/home/footer/Footer';
+import Banner from '@/components/forms/banner';
 
 export default function Layout({ children }: { children: React.ReactNode }) {
   return (
@@ -16,6 +17,7 @@ export default function Layout({ children }: { children: React.ReactNode }) {
         <Provider store={store}>
           <Providers>
             <NavigationBar />
+            <Banner />
             {children}
             <Footer />
           </Providers>
